fix: resolve comments in the PR
- Updated Checkbox component styles to use data attributes for checked and indeterminate states.
- Refactored DataTable component to improve styling and functionality, including updated row selection actions and pagination controls.
- Removed deprecated Sonner copy component and replaced it with a new Sonner implementation.
- Introduced SandboxToastPageClient for testing toast notifications with various actions.
- Created shared DataTable component for better reusability across the application.
